Update AprilTagAligner.py
For a variety of reasons the AprilTagAligner hasn't been working... I was assigning the wrong AprilTag corner locations based on the getCorner() documentation. And the conversion from robot to field orientation wasn't working out correctly on paper... and explains why the robot seemed to align well when facing 0 degrees... but went way out of alignment when facing left or right. Hopefully should work well now.